+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] - 2021-03-02
+
+### Changed
+
+- Beaker now accepts `defaultValues` as an array
+- Internal refactoring to improve group handling
+- Removed global state stores for initial values, now only generated internally for reset methods
+
+### Fixed
+
+- `formValidity` in Beaker stores is now an array
+- Group stores no longer emit twice on changes
+
 ## [0.8.7] - 2021-02-23
 
 ### Changed